fix(segmentation): Added missing tool config for labelmap and contour segmentation toolboxes.
